Programming Assignment 2: Lexical Scoping
Second programming assignment will require you to write an R function is
able to cache potentially time-consuming computations. For example,
taking the mean of a numeric vector is typically a fast operation.
However, for a very long vector, it may take too long to compute the
mean, especially if it has to be computed repeatedly (e.g. in a loop).